Developer note: download dir will not be deleted if 'CreateDownloadSubDir=false'
svn-rev: 70049

+++ src/com/exactprosystems/remotehand/web/WebUtils.java
@@ -43,6 +43,9 @@ public class WebUtils {
 
 	public static void deleteDownloadDirectory(File downloadDir)
 	{
+		if (!((WebConfiguration) Configuration.getInstance()).isCreateDownloadSubDir())
+			return;
+		
 		File[] tmp;
 		if (downloadDir != null && (tmp = downloadDir.listFiles()) != null && tmp.length == 0)
 			downloadDir.delete();
